# backend/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  models.GameSystem:
    properties:
      code:
        type: string
      createdAt:
        type: integer
      description:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      modifiedAt:
        type: integer
      name:
        type: string
    type: object
...
  models.SkillImprovementCost:
    properties:
      current_level:
        description: Aktueller Fertigkeitswert
        type: integer
      id:
        type: integer
      skillCategoryId:
        description: SkillCategoryID
        type: integer
      skillDifficultyId:
        description: SkillDifficultyID
        type: integer
      te_required:
        description: Benötigte Trainingseinheiten
        type: integer
    type: object
...
  models.Source:
    properties:
      code:
        description: z.B. "KOD", "ARK", "MYS"
        type: string
      description:
        description: Beschreibung des Werks
        type: string
      edition:
        description: z.B. "5. Edition"
        type: string
      full_name:
        description: z.B. "Midgard Regelwerk - Kodex"
        type: string
      game_system:
        type: string
      game_system_id:
        type: integer
      id:
        type: integer
      is_active:
        description: Ist das Werk aktiv/verfügbar?
        type: boolean
      is_core:
        description: Ist es ein Grundregelwerk?
        type: boolean
      name:
        description: z.B. "Kodex", "Arkanum", "Mysterium"
        type: string
      publish_year:
        description: Erscheinungsjahr
        type: integer
      publisher:
        description: z.B. "Pegasus Spiele"
        type: string
    type: object
...
host: localhost:8180
info:
  contact:
    email: bamort.support@trokan.de
    name: BaMoRT Support
  description: |-
    BaMoRT (MOAM Replacement) - Role-playing Game Character Management System API

    This API provides comprehensive character management, import/export functionality,
    master data management, and PDF generation for tabletop role-playing games.

    ## Authentication
    Most endpoints require authentication via JWT token. Obtain a token by calling POST /login.
    Include the token in the Authorization header: `Authorization: Bearer <token>`

    ## Character Import/Export
    The API supports pluggable adapters for importing/exporting characters from various systems.
    Supported formats include MOAM VTT, Foundry VTT, and more via microservice adapters.
  license:
    name: Custom License
    url: https://github.com/Bardioc26/bamort/blob/main/LICENSE
  title: BaMoRT API
  version: "2.0"
paths:
...
  /api/maintenance/game-systems:
    get:
      description: Returns list of all game systems (maintainer only)
      produces:
      - application/json
      responses:
        "200":
          description: List of game systems
          schema:
            items:
              $ref: '#/definitions/models.GameSystem'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - maintainer access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get game systems
      tags:
      - Maintenance
  /api/maintenance/game-systems/{id}:
    put:
      consumes:
      - application/json
      description: Updates an existing game system (maintainer only)
      parameters:
      - description: Game system ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated game system data
        in: body
        name: system
        required: true
        schema:
          $ref: '#/definitions/models.GameSystem'
      produces:
      - application/json
      responses:
        "200":
          description: Updated game system
          schema:
            $ref: '#/definitions/models.GameSystem'
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - maintainer access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update game system
      tags:
      - Maintenance
...
  /api/maintenance/gsm-lit-sources:
    get:
      description: Returns list of all literature sources (maintainer only)
      produces:
      - application/json
      responses:
        "200":
          description: List of literature sources
          schema:
            items:
              $ref: '#/definitions/models.Source'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - maintainer access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get literature sources
      tags:
      - Maintenance
  /api/maintenance/gsm-lit-sources/{id}:
    put:
      consumes:
      - application/json
      description: Updates an existing literature source (maintainer only)
      parameters:
      - description: Literature source ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated literature source data
        in: body
        name: source
        required: true
        schema:
          $ref: '#/definitions/models.Source'
      produces:
      - application/json
      responses:
        "200":
          description: Updated literature source
          schema:
            $ref: '#/definitions/models.Source'
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - maintainer access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update literature source
      tags:
      - Maintenance
  /api/maintenance/gsm-misc:
    get:
      description: Returns list of all miscellaneous master data entries (maintainer
        only)
      produces:
      - application/json
      responses:
        "200":
          description: List of miscellaneous data
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - maintainer access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get miscellaneous master data
      tags:
      - Maintenance
  /api/maintenance/gsm-misc/{id}:
    put:
      consumes:
      - application/json
      description: Updates a miscellaneous master data entry (maintainer only)
      parameters:
      - description: Misc data ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated miscellaneous data
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated miscellaneous data
          schema:
            type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - maintainer access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update miscellaneous master data
      tags:
      - Maintenance
...
  /api/maintenance/skill-improvement-cost2:
    get:
      description: Returns skill improvement cost table (maintainer only)
      produces:
      - application/json
      responses:
        "200":
          description: Skill improvement costs
          schema:
            items:
              $ref: '#/definitions/models.SkillImprovementCost'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - maintainer access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get skill improvement costs
      tags:
      - Maintenance
  /api/maintenance/skill-improvement-cost2/{id}:
    put:
      consumes:
      - application/json
      description: Updates a skill improvement cost entry (maintainer only)
      parameters:
      - description: Cost entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated cost data
        in: body
        name: cost
        required: true
        schema:
          $ref: '#/definitions/models.SkillImprovementCost'
      produces:
      - application/json
      responses:
        "200":
          description: Updated cost entry
          schema:
            $ref: '#/definitions/models.SkillImprovementCost'
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - maintainer access required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update skill improvement cost
      tags:
      - Maintenance
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"